HOME-11 - Record alert data in logs and notify
* Separate data fetcher from REST API
